- Added the handling of normal incidence in FastReflectTransmit (so that it behaves the same way as Erik's version). - Removed debugging code in MCMLKernel. - Added preliminary support of double precision. The input file is still parsed into floats. These floats are casted to double when allocating and transfering data to the global memory. The entire kernel is in double-precision, except WEIGHT_SCALE and RNG.

+++ gpumcml/fast/gpumcml_mem.cu
@@ -40,8 +40,8 @@ int InitDCMem(SimulationStruct *sim, UINT32 A_rz_overflow)
 
   h_simparam.num_layers = sim->n_layers;  // not plus 2 here
   h_simparam.init_photon_w = sim->start_weight;
-  h_simparam.dz = sim->det.dz;
-  h_simparam.dr = sim->det.dr;
+  h_simparam.dz = (FLOAT)sim->det.dz;
+  h_simparam.dr = (FLOAT)sim->det.dr;
   h_simparam.na = sim->det.na;
   h_simparam.nz = sim->det.nz;
   h_simparam.nr = sim->det.nr;
@@ -54,18 +54,18 @@ int InitDCMem(SimulationStruct *sim, UINT32 A_rz_overflow)
 
   for (UINT32 i = 0; i < n_layers; ++i)
   {
-    h_layerspecs[i].z0 = sim->layers[i].z_min;
-    h_layerspecs[i].z1 = sim->layers[i].z_max;
-    FLOAT n1 = sim->layers[i].n;
+    h_layerspecs[i].z0 = (FLOAT)sim->layers[i].z_min;
+    h_layerspecs[i].z1 = (FLOAT)sim->layers[i].z_max;
+    FLOAT n1 = (FLOAT)sim->layers[i].n;
     h_layerspecs[i].n = n1;
 
     // TODO: sim->layer should not do any pre-computation.
-    FLOAT rmuas = sim->layers[i].mutr;
+    FLOAT rmuas = (FLOAT)sim->layers[i].mutr;
     h_layerspecs[i].muas = FP_ONE / rmuas;
     h_layerspecs[i].rmuas = rmuas;
-    h_layerspecs[i].mua_muas = sim->layers[i].mua * rmuas;
+    h_layerspecs[i].mua_muas = (FLOAT)sim->layers[i].mua * rmuas;
 
-    h_layerspecs[i].g = sim->layers[i].g;
+    h_layerspecs[i].g = (FLOAT)sim->layers[i].g;
 
     if (i == 0 || i == n_layers-1)
     {
@@ -74,10 +74,10 @@ int InitDCMem(SimulationStruct *sim, UINT32 A_rz_overflow)
     }
     else
     {
-      FLOAT n2 = sim->layers[i-1].n;
+      FLOAT n2 = (FLOAT)sim->layers[i-1].n;
       h_layerspecs[i].cos_crit0 = (n1 > n2) ?
         sqrtf(FP_ONE - n2*n2/(n1*n1)) : MCML_FP_ZERO;
-      n2 = sim->layers[i+1].n;
+      n2 = (FLOAT)sim->layers[i+1].n;
       h_layerspecs[i].cos_crit1 = (n1 > n2) ?
         sqrtf(FP_ONE - n2*n2/(n1*n1)) : MCML_FP_ZERO;
     }
